+++ src/utils.ts
@@ -87,14 +87,14 @@ export function getBsonSchema<const TMeta extends _.Meta>(
 //@ts-expect-error - Vite handles this import.meta check
 if (import.meta.vitest) {
   //@ts-expect-error - Vite handles this top-level await
-  const [{ describe, it, expect }, { default: B }] = await Promise.all([
+  const [{ describe, it, expect }, { default: b }] = await Promise.all([
     import("vitest"),
     import("src")
   ]);
 
   describe("getBsonSchema", () => {
     it("should return the correct schema for a union", () => {
-      const borg = B.union([B.string(), B.number().nullable(), B.boolean()]);
+      const borg = b.union([b.string(), b.number().nullable(), b.boolean()]);
 
       const schema = getBsonSchema(borg.meta);
       expect(schema).toEqual({
@@ -126,7 +126,7 @@ if (import.meta.vitest) {
     });
 
     it("should return the correct schema for a string", () => {
-      const borg = B.string().minLength(5).maxLength(10);
+      const borg = b.string().minLength(5).maxLength(10);
 
       const schema = getBsonSchema(borg.meta);
       expect(schema).toEqual({
@@ -152,7 +152,7 @@ if (import.meta.vitest) {
     });
 
     it("should return the correct schema for a number", () => {
-      const borg = B.number().min(5).max(10);
+      const borg = b.number().min(5).max(10);
 
       const schema = getBsonSchema(borg.meta);
       expect(schema).toEqual({
@@ -170,7 +170,7 @@ if (import.meta.vitest) {
     });
 
     it("should return the correct schema for a boolean", () => {
-      const borg = B.boolean();
+      const borg = b.boolean();
 
       const schema = getBsonSchema(borg.meta);
       expect(schema).toEqual({
@@ -184,7 +184,7 @@ if (import.meta.vitest) {
     });
 
     it("should return the correct schema for an id", () => {
-      const borg = B.id();
+      const borg = b.id();
 
       const schema = getBsonSchema(borg.meta);
       expect(schema).toEqual({
@@ -198,7 +198,7 @@ if (import.meta.vitest) {
     });
 
     it("should return the correct schema for an array", () => {
-      const borg = B.array(B.string().minLength(5).maxLength(10))
+      const borg = b.array(b.string().minLength(5).maxLength(10))
         .minItems(1)
         .maxItems(5);
       const schema = getBsonSchema(borg.meta);
@@ -227,22 +227,22 @@ if (import.meta.vitest) {
     });
 
     it("should return the correct schema for an object", () => {
-      const borg = B.object({
-        a: B.string().minLength(5).maxLength(10),
-        b: B.number().min(5).max(10),
-        c: B.boolean(),
-        d: B.id(),
-        e: B.array(B.string().minLength(5).maxLength(10))
+      const borg = b.object({
+        a: b.string().minLength(5).maxLength(10),
+        b: b.number().min(5).max(10),
+        c: b.boolean(),
+        d: b.id(),
+        e: b.array(b.string().minLength(5).maxLength(10))
           .minItems(1)
           .maxItems(5),
-        f: B.object({
-          g: B.string().minLength(5).maxLength(10),
-          h: B.number().min(5).max(10).optional(),
-          i: B.object({}).additionalProperties(B.string().pattern("^a$")),
-          j: B.object({}).additionalProperties("passthrough").nullable()
+        f: b.object({
+          g: b.string().minLength(5).maxLength(10),
+          h: b.number().min(5).max(10).optional(),
+          i: b.object({}).additionalProperties(b.string().pattern("^a$")),
+          j: b.object({}).additionalProperties("passthrough").nullable()
         }).additionalProperties("strip"),
-        k: B.string().optional(),
-        l: B.union([B.string(), B.number().nullable(), B.boolean()])
+        k: b.string().optional(),
+        l: b.union([b.string(), b.number().nullable(), b.boolean()])
       }).additionalProperties("strict");
 
       const schema = getBsonSchema(borg.meta);